tensorflow v1 inside loop
