+++ public/assets/AIDecision.json
@@ -1,0 +1,1 @@
+[{"BuyProb":0.7,"HoldProb":49.5,"SellProb":49.8,"dateCreated":"1565171840","CurrentPrice":11629.93486063,"AIDecision":"Sell","SellIfPrice":11629.93486063,"BuyIfPrice":11585.857213329973},{"BuyProb":0.3,"HoldProb":56.4,"SellProb":43.3,"dateCreated":"1565171840","CurrentPrice":11629.93486063,"AIDecision":"Hold","SellIfPrice":11652.19002198484,"BuyIfPrice":11585.857213329973},{"BuyProb":0,"HoldProb":56.7,"SellProb":43.3,"dateCreated":"1565171840","CurrentPrice":11629.93486063,"AIDecision":"Hold","SellIfPrice":11652.19002198484,"BuyIfPrice":11585.857213329973},{"BuyProb":0.2,"HoldProb":95.8,"SellProb":4,"dateCreated":"1565190383","CurrentPrice":11660.00000002,"AIDecision":"Hold","SellIfPrice":11682.297982033251,"BuyIfPrice":11612.28572707445}]
